refactor: extract pure functions from app.js into lib.js (ES modules)
Extract ~180 lines of pure computation (LTTB downsampling, bucketize,
histogram, quality score, stats helpers, bsearch, filterRange, rgba)
into a dedicated ES module gh-pages/lib.js.

app.js now imports from lib.js and serves as the DOM orchestrator.
HTML template updated from <script defer> to <script type="module">.

Closes #26
