+++ gh-pages/index.template.html
@@ -114,7 +114,7 @@
 
     <script defer src="https://cdn.jsdelivr.net/npm/chart.js@4"></script>
     <script defer src="https://cdn.jsdelivr.net/npm/chartjs-adapter-date-fns@3"></script>
-    <script defer src="app.js"></script>
+    <script type="module" src="app.js"></script>
     <script>
       setTimeout(function () {
         location.reload();
